+++ frameworks/react/src/toast/stories/toast.stories.tsx
@@ -7,99 +7,95 @@ const meta: Meta = {
 }
 export default meta
 
-export const Basic = () => {
-  const [Toaster, toast] = createToaster({
-    placement: 'top-end',
-    render(toast) {
-      return (
-        <Toast.Root>
-          <Toast.Title>{toast.title}</Toast.Title>
-          <Toast.Description>{toast.description}</Toast.Description>
-          <Toast.CloseTrigger>Close</Toast.CloseTrigger>
-        </Toast.Root>
-      )
-    },
-  })
+const [Toaster, toast] = createToaster({
+  placement: 'top-end',
+})
 
+export const Basic = () => {
   const handleToast = () => {
-    const id = toast.create({ title: 'Title', description: 'Description' })
-    console.log('Toast ID:', id)
+    toast.create({ id: '1', title: 'Title', description: 'Description' })
   }
 
   return (
     <>
       <button onClick={handleToast}>Toast</button>
-      <Toaster />
-    </>
-  )
-}
-
-export const Customized = () => {
-  const [Toaster, toast] = createToaster({
-    placement: 'bottom-start',
-    render(toast) {
-      return (
-        <Toast.Root>
-          <Toast.Title>{toast.title}</Toast.Title>
-          <Toast.Description>{toast.description}</Toast.Description>
-          <Toast.CloseTrigger>Close</Toast.CloseTrigger>
-        </Toast.Root>
-      )
-    },
-  })
-
-  return (
-    <>
-      <button
-        onClick={() =>
-          toast.create({
-            title: 'Success',
-            description: 'This is a success toast',
-            type: 'success',
-            duration: 20000,
-            removeDelay: 250,
-          })
+      <Toaster>
+        {(toasts) =>
+          toasts.map((toast) => (
+            <Toast.Root key={toast.id} toast={toast}>
+              <Toast.Title>{toast.title}</Toast.Title>
+              <Toast.Description>{toast.description}</Toast.Description>
+              <Toast.CloseTrigger>Close</Toast.CloseTrigger>
+            </Toast.Root>
+          ))
         }
-      >
-        Toast
-      </button>
-      <Toaster />
+      </Toaster>
     </>
   )
 }
 
-export const CustomRender = () => {
-  const [Toaster, toast] = createToaster({
-    placement: 'top-end',
-    // custom render may go directly into the function below
-    render(toast) {
-      return (
-        <Toast.Root>
-          <Toast.Title>{toast.title}</Toast.Title>
-          <Toast.Description>{toast.description}</Toast.Description>
-          <Toast.CloseTrigger>Close</Toast.CloseTrigger>
-        </Toast.Root>
-      )
-    },
-  })
+// export const Customized = () => {
+//   const [Toaster, toast] = createToaster({
+//     placement: 'bottom-start',
+//     render(toast) {
+//       return (
 
-  return (
-    <>
-      <button
-        onClick={() =>
-          toast.create({
-            title: 'Please checkout',
-            render: (toast) => (
-              <div>
-                {toast.title} <a href="https://ark-ui.com">Ark UI</a>
-              </div>
-            ),
-          })
-        }
-      >
-        Toast
-      </button>
-      <Toaster />
-    </>
-  )
-}
+//       )
+//     },
+//   })
+
+//   return (
+//     <>
+//       <button
+//         onClick={() =>
+//           toast.create({
+//             title: 'Success',
+//             description: 'This is a success toast',
+//             type: 'success',
+//             duration: 20000,
+//             removeDelay: 250,
+//           })
+//         }
+//       >
+//         Toast
+//       </button>
+//       <Toaster />
+//     </>
+//   )
+// }
+
+// export const CustomRender = () => {
+//   const [Toaster, toast] = createToaster({
+//     placement: 'top-end',
+//     // custom render may go directly into the function below
+//     render(toast) {
+//       return (
+//         <Toast.Root>
+//           <Toast.Title>{toast.title}</Toast.Title>
+//           <Toast.Description>{toast.description}</Toast.Description>
+//           <Toast.CloseTrigger>Close</Toast.CloseTrigger>
+//         </Toast.Root>
+//       )
+//     },
+//   })
+
+//   return (
+//     <>
+//       <button
+//         onClick={() =>
+//           toast.create({
+//             title: 'Please checkout',
+//             render: (toast) => (
+//               <div>
+//                 {toast.title} <a href="https://ark-ui.com">Ark UI</a>
+//               </div>
+//             ),
+//           })
+//         }
+//       >
+//         Toast
+//       </button>
+//       <Toaster />
+//     </>
+//   )
+// }