+++ frameworks/react/src/toast/toast-title.tsx
@@ -1,13 +1,13 @@
 import { mergeProps } from '@zag-js/react'
 import { forwardRef } from 'react'
 import { ark, type HTMLArkProps } from '../factory'
-import { useToastContext } from './use-toast-context'
+import { useToasterItemContext } from './use-toaster-item-context'
 
 export interface ToastTitleProps extends HTMLArkProps<'div'> {}
 
 export const ToastTitle = forwardRef<HTMLDivElement, ToastTitleProps>((props, ref) => {
-  const toast = useToastContext()
-  const mergedProps = mergeProps(toast.titleProps, props)
+  const toasterItem = useToasterItemContext()
+  const mergedProps = mergeProps(toasterItem.titleProps, props)
 
   return <ark.div {...mergedProps} ref={ref} />
 })